Move tests to same-name package in different source directory
