+++ test/speciesnetwork/AllTests.java
@@ -1,4 +1,4 @@
-package snetworktests;
+package speciesnetwork;
 
 import org.junit.runner.RunWith;
 import org.junit.runners.Suite;
