+++ test/speciesnetwork/BirthHybridizationTest.java
@@ -1,4 +1,4 @@
-package snetworktests;
+package speciesnetwork;
 
 import org.junit.Test;
 import static org.junit.Assert.assertEquals;
